Split node_primary_addresses into a separate module
